refactoring of subcommands.
