+++ rpc-client/src/subcommands/mod.rs
@@ -1,0 +1,14 @@
+pub use accounts_subcommands::AccountCommand;
+pub use accounts_subcommands::HandleSubcommand;
+pub use blockchain_subcommands::BlockchainCommand;
+pub use mempool_subcommands::MempoolCommand;
+pub use network_subcommands::NetworkCommand;
+pub use transactions_subcommands::TransactionCommand;
+pub use validator_subcommands::ValidatorCommand;
+
+mod accounts_subcommands;
+mod blockchain_subcommands;
+mod mempool_subcommands;
+mod network_subcommands;
+mod transactions_subcommands;
+mod validator_subcommands;
